# agent-manager-service/docs/api_v1_openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Agent Manager Service API
servers:
  - url: /api/v1/agent-manager-service
paths:
  /orgs:
    get:
      summary: List all organizations
      operationId: listOrganizations
      parameters:
        - name: limit
          in: query
          description: Maximum number of results to return
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - name: offset
          in: query
          description: Number of results to skip
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationListResponse"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create a new organization
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "202":
          description: Organization creation initiated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orgs/{orgName}:
    get:
      summary: Get organization details
      operationId: getOrganization
      parameters:
        - name: orgName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Organization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationResponse"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orgs/{orgName}/projects:
    post:
      summary: Create a new project
      operationId: createProject
      parameters:
        - name: orgName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "202":
          description: Project creation initiated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Project already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      summary: List all projects in an organization
      operationId: listProjects
      parameters:
        - name: orgName
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of results to return
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - name: offset
          in: query
          description: Number of results to skip
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orgs/{orgName}/projects/{projName}:
    get:
      summary: Get project details
      operationId: getProject
      parameters:
        - name: orgName
          in: path
          required: true
          schema:
            type: string
        - name: projName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orgs/{orgName}/projects/{projName}/agents:
    post:
      summary: Create a new agent
      operationId: createAgent
      parameters:
        - name: orgName
          in: path
          required: true
          schema:
            type: string
        - name: projName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRequest"
      responses:
        "202":
          description: Agent creation initiated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Agent already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      summary: List all agents in a project of an organization
      operationId: listAgents
      parameters:
        - name: orgName
          in: path
          required: true
          schema:
            type: string
        - name: projName
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of results to return
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - name: offset
          in: query
          description: Number of results to skip
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentListResponse"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    CreateOrganizationRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Name of the organization
    CreateProjectRequest:
      type: object
      required:
        - name
        - deploymentPipeline
        - displayName
      properties:
        name:
          type: string
          description: Name of the project
        displayName:
          type: string
          description: Display name of the project
        description:
          type: string
          description: Description of the project
        deploymentPipeline:
          type: string
    Provisioning:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum: [internal]
          description: Type of provisioning
        repository:
          $ref: "#/components/schemas/RepositoryConfig"
    CreateAgentRequest:
      type: object
      required:
        - name
        - displayName
        - provisioning
      properties:
        provisioning:
          $ref: "#/components/schemas/Provisioning"
        name:
          type: string
          description: Unique name of the agent
        displayName:
          type: string
          description: Display name of the agent
        description:
          type: string
          description: Description of the agent
        runtimeConfigs:
          $ref: "#/components/schemas/RuntimeConfiguration"
        inputInterface:
          $ref: "#/components/schemas/InputInterface"
    AgentResponse:
      type: object
      properties:
        name:
          type: string
        displayName:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        projectName:
          type: string
        status:
          type: string
          description: Current status of the agent
        provisioning:
          $ref: "#/components/schemas/Provisioning"
      required:
        - name
        - displayName
        - createdAt
        - projectName
        - description
        - provisioning
...
    AgentListResponse:
      type: object
      properties:
        agents:
          type: array
          items:
            $ref: "#/components/schemas/AgentResponse"
        total:
          type: integer
          description: Total number of agents matching the query
        limit:
          type: integer
        offset:
          type: integer
      required:
        - agents
        - total
        - limit
        - offset
...
    RepositoryConfig:
      type: object
      description: Git repository configuration
      required:
        - url
        - branch
        - appPath
      properties:
        url:
          type: string
          description: URL of the git repository
        branch:
          type: string
          description: Branch to use
        appPath:
          type: string
          description: Path within the repository where the source code resides
...
    RuntimeConfiguration:
      type: object
      properties:
        env:
          type: array
          description: Environment variables
          items:
            $ref: "#/components/schemas/EnvironmentVariable"
        runCommand:
          type: string
        languageVersion:
          type: string
        language:
          type: string
      required:
        - language
        - languageVersion
    EnvironmentVariable:
      type: object
      required:
        - key
        - value
      properties:
        key:
          type: string
        value:
          type: string
    InputInterface:
      type: object
      description: Endpoint configurations
      required:
        - type
      properties:
        type:
          type: string
          description: Type of input interface (e.g., default, custom)
          enum: [DEFAULT, CUSTOM]
        customOpenAPISpec:
          $ref: "#/components/schemas/EndpointSpec"
    EndpointSpec:
      type: object
      required:
        - port
        - schema
        - basePath
      properties:
        port:
          type: integer
          description: Port number
          minimum: 1
          maximum: 65535
        schema:
          $ref: "#/components/schemas/EndpointSchema"
        basePath:
          type: string
          description: Base path for the endpoint
    EndpointSchema:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          description: Schema content
...
    OrganizationResponse:
      type: object
      properties:
        name:
          type: string
          description: Name of the organization
        displayName:
          type: string
          description: Display name of the organization
        description:
          type: string
          description: Description of the organization
        namespace:
          type: string
          description: Namespace of the organization
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the organization was created
      required:
        - name
        - createdAt
        - displayName
        - description
        - namespace
    OrganizationListItem:
      type: object
      properties:
        name:
          type: string
          description: Name of the organization
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the organization was created
      required:
        - name
        - createdAt
    OrganizationListResponse:
      type: object
      properties:
        organizations:
          type: array
          items:
            $ref: "#/components/schemas/OrganizationListItem"
        total:
          type: integer
          description: Total number of organizations matching the query
        limit:
          type: integer
        offset:
          type: integer
      required:
        - organizations
        - total
        - limit
        - offset
    ProjectResponse:
      type: object
      properties:
        name:
          type: string
          description: Name of the project
        orgName:
          type: string
          description: Name of the organization
        displayName:
          type: string
          description: Display name of the project
        description:
          type: string
          description: Description of the project
        deploymentPipeline:
          type: string
          description: Deployment pipeline associated with the project
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the project was created
      required:
        - name
        - orgName
        - displayName
        - description
        - deploymentPipeline
        - createdAt
    ProjectListItem:
      type: object
      properties:
        name:
          type: string
          description: Name of the project
        orgName:
          type: string
          description: Name of the organization
        displayName:
          type: string
          description: Display name of the project
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the project was created
      required:
        - name
        - orgName
        - displayName
        - createdAt
    ProjectListResponse:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/ProjectListItem"
        total:
          type: integer
          description: Total number of projects matching the query
        limit:
          type: integer
        offset:
          type: integer
      required:
        - projects
        - total
        - limit
        - offset
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Human-readable error message
        description:
          type: string
          description: Error description
        additionalData:
          type: object
          additionalProperties: true
      required:
        - message